+++ Changes.md
@@ -1,6 +1,7 @@
 HEAD
 -----------
 - Improved exception handling in NodeWatcher.
+- Stubbed Client#client to return itself, fixes a fork reconnect bug with Resque (dbalatero)
 
 0.9.1
 -----------