+++ Changes.md
@@ -1,3 +1,7 @@
+0.9.7
+-----------
+- Stubbed Client#client to return itself, fixes a fork reconnect bug with Resque (dbalatero)
+
 0.9.6
 -----------
 - Handle the node discovery error condition where the znode points to a master that is now a slave.
@@ -19,7 +23,6 @@
 0.9.2
 -----------
 - Improved exception handling in NodeWatcher.
-- Stubbed Client#client to return itself, fixes a fork reconnect bug with Resque (dbalatero)
 
 0.9.1
 -----------